// src/payload-types.ts
export interface User {
  id: string;
  sub?: string | null;
  username: string;
  name: string;
  first_name: string;
  last_name: string;
  position?: string | null;
  image?: (string | null) | Media;
  description?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  mobile?: string | null;
  phone?: string | null;
  gender?: ('male' | 'female') | null;
  roles: ('admin' | 'user')[];
  Pipelines?: {
    docs?: (string | Pipeline)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  password?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: string;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "pipelines".
 */
export interface Pipeline {
  id: string;
  title: string;
  description?: string | null;
  backgroundColor?: string | null;
  owner: string | User;
  members?: (string | User)[] | null;
  isPrivate?: boolean | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "stages".
 */
export interface Stage {
  id: string;
  title: string;
  pipeline: string | Pipeline;
  position: number;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "tasks".
 */
export interface Task {
  id: string;
  title: string;
  description?: string | null;
  stage: string | Stage;
  position: number;
  assignee?: (string | null) | User;
  observers?: (string | User)[] | null;
  dueDate?: string | null;
  labels?:
    | {
        name?: string | null;
        color?: string | null;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}